backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
python-dotenv
alembic
psycopg2-binary
python-multipart
pyjwt
passlib[bcrypt]
google-generativeai
